rename gooq binary from cmd to gooq
