+++ cmd/gooq/generator/generator.go
