+++ cmd/gooq/generator/root.go
