+++ cmd/gooq/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/lumina-tech/gooq/cmd/generator"
+	"github.com/lumina-tech/gooq/cmd/gooq/generator"
 )
 
 func main() {
